pyproject.toml
[tool.poetry]
name = "seazone"
version = "0.1.0"
description = ""
authors = ["Carlos Cezar Negreiros <cezarnegreiros07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
pydantic-settings = "^2.6.0"
asyncpg = "^0.30.0"
ruff = "^0.7.0"
taskipy = "^1.13.0"
pytest-cov = "^5.0.0"
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.24.0"
pytest = "^8.3.3"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations', '.env', '.venv', '.idea']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'